Fix subscription filtering for GQL introspection
